requirements.txt
llama-index
llama-index-vector-stores-chroma
openai
chromadb
fastapi
uvicorn
python-dotenv